refactor: remove cloud condition from optimize
