--- optimize/util/optimize-commons/src/main/java/io/camunda/optimize/service/util/configuration/condition/CamundaCloudCondition.java
@@ -1,21 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.optimize.service.util.configuration.condition;
-
-import org.springframework.context.annotation.Condition;
-import org.springframework.context.annotation.ConditionContext;
-import org.springframework.core.type.AnnotatedTypeMetadata;
-
-// This condition is enabled if either Cloud SaaS or CCSM conditions are enabled
-public class CamundaCloudCondition implements Condition {
-  @Override
-  public boolean matches(final ConditionContext context, final AnnotatedTypeMetadata metadata) {
-    return new CCSaaSCondition().matches(context, metadata)
-        || new CCSMCondition().matches(context, metadata);
-  }
-}
